Move commcare_cloud alias to its own file
to avoid import cycles
