+++ src/commcare_cloud/commcare_cloud.py
@@ -8,7 +8,6 @@ import sys
 import warnings
 from collections import OrderedDict
 
-import six
 from clint.textui import puts, colored
 
 from commcare_cloud.cli_utils import print_command
@@ -157,32 +156,6 @@ def make_command_parser(available_envs, formatter_class=RawTextHelpFormatter,
         for alias in cmd.aliases:
             commands[alias] = cmd
     return parser, subparsers, commands
-
-
-def _encode_args(*args, **kwargs):
-    argv = []
-
-    def encode_string(value):
-        if isinstance(value, six.string_types + six.integer_types):
-            return unicode(value).encode('utf-8')
-        else:
-            TypeError("Do not know how to interpret type {} as a command-line argument: {}"
-                      .format(type(value), value))
-    for arg in args:
-        argv.append(unicode(arg).encode('utf-8'))
-    for key, value in kwargs.items():
-        if value is False:
-            continue
-        elif value is True:
-            argv.append('--{}'.format(key))
-        else:
-            argv.extend(['--{}'.format(key), encode_string(value)])
-    return argv
-
-
-def commcare_cloud(*args, **kwargs):
-    argv = _encode_args('commcare-cloud', *args, **kwargs)
-    main(argv)
 
 
 def main(input_argv=sys.argv):
